+++ geopackage-sdk/src/androidTest/java/mil/nga/geopackage/test/extension/index/FeatureTableIndexCreateTest.java
@@ -29,4 +29,9 @@ public class FeatureTableIndexCreateTest extends CreateGeoPackageTestCase {
 
     }
 
+    @Override
+    public boolean allowEmptyFeatures() {
+        return false;
+    }
+
 }